Finalizing Parser

+++ src/main/java/com/eeg/xmlfootballparser/models/Team.java
@@ -8,6 +8,9 @@ import lombok.ToString;
 import java.util.HashMap;
 import java.util.List;
 
+/**
+ * JAVA POJO (Team). Utilizes lombok to avoid common boilerplate code.
+ * */
 @Builder
 @Getter
 @ToString
@@ -16,5 +19,5 @@ public class Team<P extends Player> {
     private short score;
     private Side side;
     private List<P> players;
-    private HashMap<String, Short> statistics;
+    private HashMap<String, Float> statistics;
 }
